Shortcut get_revids_from when we aren't actually checking merges.
The code is there to handle stuff like 'show mainline revs that merged these file
changes'. However, it is way overkill when all you need is the simple mainline.

This avoids scaling something that is a simple linear load into a O(history) operation.

Saves ~30ms on emacs. (Down to 215ms for emacs/trunk/changes).