Add a <noop> command to the RequestWorker.
This allows us to push the workers to stop immediately, in case
they are currently blocked waiting on another item in the queue.
This makes the test suite integration tests faster, but also
makes the script runner exit in a timely manner as well.

Also creating a trivial load_test script, just for example
purposes.